+++ R/results-widecr.R
@@ -64,10 +64,12 @@
 #' throws error otherwise.
 #'
 #' @return `is_widecr()` returns `TRUE` if its argument is appropriate object
-#'   of class `widecr`: it should inherit classes `widecr`, `tbl_df` (in other
-#'   words, to be [tibble][tibble::tibble]) and have complete pairs of
-#'   "player"-"score" columns where pair is detected by __digits__ after strings
-#'   "player" and "score" respectively.
+#' of class `widecr`: it should inherit classes `widecr`, `tbl_df` (in other
+#' words, to be [tibble][tibble::tibble]) and have complete pairs of
+#' "player"-"score" columns where pair is detected by __digits__ after strings
+#' "player" and "score" respectively. Columns of "player" and "score" types
+#' shouldn't have any extra symbols except type name and digits after it. All
+#' other columns are considered as "extra columns".
 #'
 #' `as_widecr()` returns an object of class `widecr`.
 #'
@@ -100,7 +102,7 @@ is_widecr <- function(cr_data) {
   }
   names_cr <- tolower(colnames(cr_data))
   names_df <- tibble::tibble(
-    name = names_cr[grepl("player|score", x = names_cr)]
+    name = names_cr[grepl("^(player|score)[0-9]+$", x = names_cr)]
   )
 
   if (nrow(names_df) == 0) {